# backend/documentation/openapi.yaml
openapi: 3.1.0
info:
  title: D-Bo API Documentation
  version: 0.1.0
  description:
    This is the backend REST API for the application **D-Bo** - empowering users to play a familiar
    card game with friends and family together in real time.
    <br /><br />
    The REST API handles all stateless HTTP requests, including player authentication and account
    management, matchmaking, and game session lifecycle operations.
    <br /><br />
    To find out more about how **real-time player-to-player interaction** happens, check out the
    [Async API](./asyncapi.yaml) documentation.
    <hr />
servers:
  - url: https://api.d-bo.bigdevdog.com
    description: Production server
  - url: http://localhost:60600
    description: Local test server
...
paths:
...
  /players/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh the player's authentication tokens
      description: |
        This endpoint is used to generate a fresh access token for the player to use to
        authenticate themselves for 15 minutes, and replaces their old refresh token with a fresh
        one in the database.

        This request requires that an HTTP-only cookie be set for this path, containing the refresh
        token id and secret. The cookie will be extracted, and then its credentials will be verified
        against the existing refresh token. The refresh token will not work if it is more than 30 
        days old, or if it has been marked as "revoked". It will then find the associated player
        account in order to generate new authentication tokens for the player.
      parameters:
        - in: cookie
          name: refresh_token
          required: true
          description: The refresh token header
          schema:
            type: string
            example: 1f1cbaac-34d6-41d7-ba73-5137831a27fb:aGw_YnVbBT8WOkYRyWA1AqhYa0Hvmz7Jwe4tIMppPnc
      responses:
        200:
          description: The request was successful. A new cookie has been set, good for 30 days, and
            the body contains the new access token to authenticate the player for 15 minutes.
          headers:
            Set-Cookie:
              description:
                An HTTP-only cookie which can be used to refresh an expired access token.
              schema:
                type: string
                example: 
                  refresh_token=6727bdde-ecf7-4f2c-9f82-fa1fdcb61ed1:JroD-BB9RLXWqB7A-7BNQT6kcNkz-xzLMSY9NonP4_Y; 
                  HttpOnly; SameSite=Strict; Path=/players/refresh
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: 
                      eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhMjljMDZhZS05MWM4LTQ4NDktYmM3MS05NmIwYmNlOTAzMDgiLCJleHAiOjE3NTczNzU0MjAsImlhdCI6MTc1NzM3NDUyMH0.2lQnOVkusYC0nfue8FbTQjFxM9fXojdTGEvG8djAAx4
        401:
          description: The request failed because the cookie was not set, was formatted incorrectly, 
            did not match the records in the database, or was not associated with an existing player
            account.
        403:
          description: The request failed because the token had been revoked.
        410:
          description: The request failed because the token was expired after more than 30 days.
        500:
          $ref: '#/components/responses/ServerSideError'
components:
...
  responses:
    ServerSideError:
      description: The request could not be completed due to a server-side error.
...